requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
loguru==0.7.2
nest-asyncio==1.5.8
